% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsic.R
\name{g}
\alias{g}
\title{Shortcut for glue::glue extrapolating character strings}
\usage{
g(str, .envir = parent.frame())
}
\arguments{
\item{str}{a character string possibly containing expressions in \{...\}}

\item{.envir}{environment where to evaluate expressions}
}
\value{
a character string with expressions evaluated and numbers formatted
}
\description{
Shortcut for glue::glue extrapolating character strings
}
